tests: reorganize, deduplicate, and de-bloat the suite
Streamline the manually-run test suite without changing coverage
(898 tests collected before and after; all pass).

Structure:
- Combine test_astro/test_kpf_utils/test_stats into test_utils.py.
- Split test_data_models.py into test_data_models_l{0,1,2,4}.py, mirroring
  the test_quicklook_l*.py per-level layout.

Consolidation:
- Add tests/conftest.py holding the synthetic-FITS fixtures that were
  duplicated across the data-model level modules, seeded for determinism,
  plus a shared image_hdu() helper.
- Extract the byte-identical make_l1_arrays() from test_master_bias/dark
  into tests/_masters.py.

Config / determinism:
- Add [tool.pytest.ini_options] (testpaths + filterwarnings for the known
  astropy VerifyWarning) and register a requires_testdata marker via
  conftest, converting the existing testdata skipif to it.
- Seed remaining unseeded np.random usages.

De-bloat:
- Parametrize copy-pasted shape/dtype assertion families across the master,
  image-assembly, spectral-extraction, science-recipe, and utils tests.
