+++ tests/test_data_models_l0.py
@@ -1,0 +1,70 @@
+"""
+Tests for the KPF0 (raw CCD / L0) data model.
+
+Uses synthetic FITS fixtures — no real KPF data needed.
+"""
+
+import numpy as np
+import pytest
+
+from kpfpipe.data_models.level0 import KPF0
+
+# synthetic_l0_file and synthetic_l0_minimal fixtures live in tests/conftest.py
+
+
+class TestKPF0:
+    def test_from_fits(self, synthetic_l0_file):
+        l0 = KPF0.from_fits(synthetic_l0_file)
+        assert l0.level == 0
+        assert l0.obs_id == "KP.20240113.23249.10"
+        assert "GREEN_AMP1" in l0.extensions
+        assert "GREEN_AMP2" in l0.extensions
+        assert "RED_AMP1" in l0.extensions
+        assert "CA_HK" in l0.extensions
+        assert "TELEMETRY" in l0.extensions
+        assert l0.data["GREEN_AMP1"].shape == (32, 32)
+        assert l0.headers["PRIMARY"]["INSTRUME"] == "KPF"
+
+    def test_from_fits_minimal(self, synthetic_l0_minimal):
+        l0 = KPF0.from_fits(synthetic_l0_minimal)
+        assert l0.level == 0
+        assert l0.obs_id == "KP.20240113.00001.00"
+        assert "PRIMARY" in l0.extensions
+        assert len(l0.extensions) == 1
+
+    def test_round_trip(self, synthetic_l0_file, tmp_path):
+        l0 = KPF0.from_fits(synthetic_l0_file)
+        original_green = l0.data["GREEN_AMP1"].copy()
+
+        out_fn = str(tmp_path / "roundtrip_l0.fits")
+        l0.to_fits(out_fn)
+
+        l0_reread = KPF0.from_fits(out_fn)
+        np.testing.assert_array_almost_equal(
+            l0_reread.data["GREEN_AMP1"], original_green
+        )
+        assert l0_reread.headers["PRIMARY"]["INSTRUME"] == "KPF"
+
+    def test_receipt_tracking(self, synthetic_l0_file, tmp_path):
+        l0 = KPF0.from_fits(synthetic_l0_file)
+        assert len(l0.receipt) >= 1
+        assert "from_fits" in l0.receipt["Module_Name"].values
+
+        out_fn = str(tmp_path / "receipt_test.fits")
+        l0.to_fits(out_fn)
+        assert "to_fits" in l0.receipt["Module_Name"].values
+
+    def test_generate_filename(self, synthetic_l0_file):
+        l0 = KPF0.from_fits(synthetic_l0_file)
+        assert l0.generate_standard_filename() == "KP.20240113.23249.10.fits"
+
+    def test_file_not_found(self):
+        with pytest.raises(IOError):
+            KPF0.from_fits("/nonexistent/path.fits")
+
+    def test_non_fits_file(self, tmp_path):
+        fn = str(tmp_path / "not_a_fits.txt")
+        with open(fn, "w") as f:
+            f.write("hello")
+        with pytest.raises(IOError):
+            KPF0.from_fits(fn)
